refactor: ルーティングロジックをApp.tsxに移動
useContextを使用するため
